basic code

+++ src/main/java/allaboutecm/dataaccess/neo4j/Neo4jDAO.java
@@ -55,7 +55,18 @@ public class Neo4jDAO implements DAO {
 
     @Override
     public <T extends Entity> void delete(T entity) {
-        session.delete(entity);
+        Class clazz = entity.getClass();
+        T existingEntity = findExistingEntity(entity, clazz);
+        if (null == existingEntity) {
+            throw new IllegalArgumentException("the entity not exist");
+        } else if (clazz.equals(Musician.class)) {
+            Musician musician = (Musician) entity;
+            Collection<MusicianInstrument> musicianInstruments = this.loadAll(MusicianInstrument.class);
+            for (MusicianInstrument musicianInstrument : musicianInstruments) {
+                session.delete(musicianInstrument);
+            }
+            session.delete(entity);
+        }
     }
 
     @Override
